Create events and posts pages
