+++ app/page.tsx
@@ -8,9 +8,13 @@ export default function Home({}) {
       <h1>This is the home page, bruh</h1>
       <hr />
       <h2>The "Events" component goes here.</h2>
-      <Link href="/events">Click here to see all events...</Link>
+      <Link href="/events" style={{ textDecoration: 'underline' }}>
+        Click here to see all events...
+      </Link>
       <h2>The "Posts" component goes here.</h2>
-      <Link href="/posts">Click here to see all posts...</Link>
+      <Link href="/posts" style={{ textDecoration: 'underline' }}>
+        Click here to see all posts...
+      </Link>
     </main>
   )
 }
