--- app/posts/[slug]/page.tsx
@@ -1,28 +1,0 @@
-import { FC } from 'react'
-import {allDocs} from "contentlayer/generated"
-import { Mdx } from '@/components/MDX-components';
-
-interface PageProps {
-    params: {
-        slug: string
-    }
-}
-
-async function getDocFromParams(slug: string) {
-    const post = allDocs.find((post) => post.slugAsParams === slug);
-    return post;
-  }
-
-const page = async ({ params }: PageProps) => {
-    const post = await getDocFromParams(params.slug)
-    
-    if(!post) {
-        return <div>404 sorry you poor bitdev</div>
-    }
-
-    return <div>
-        <Mdx slug={post._raw.sourceFileName.slice(0, -3)} code={post.body.code} />
-    </div>
-}
-
-export default page
