+++ content/events/2023-07-29.md
@@ -1,0 +1,9 @@
+---
+date: '2023-07-29.md'
+title: 'Socratic Seminar for 2023-07-29'
+type: Doc
+---
+
+This is the body of the event for 2023-07-29.
+
+Bitcoin ipsum...
